support zip in memory

+++ ofdrw-pkg/src/main/java/org/ofdrw/pkg/container/VirtualContainer.java
@@ -5,11 +5,10 @@ import org.dom4j.DocumentException;
 import org.dom4j.Element;
 import org.ofdrw.core.DefaultElementProxy;
 import org.ofdrw.core.basicType.ST_Loc;
+import org.ofdrw.pkg.enums.ContainerType;
 import org.ofdrw.pkg.tool.ElemCup;
 
-import java.io.Closeable;
-import java.io.FileNotFoundException;
-import java.io.IOException;
+import java.io.*;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
@@ -24,6 +23,16 @@ import java.util.function.Function;
  * @since 2020-04-02 19:01:04
  */
 public class VirtualContainer implements Closeable {
+
+    /**
+     * 容器类型
+     */
+    protected ContainerType containerType;
+
+    /**
+     * ZIP内存容器
+     */
+    protected ZipContainer zipContainer;
 
     /**
      * 文件根路径(完整路径包含当前文件名)
@@ -44,7 +53,6 @@ public class VirtualContainer implements Closeable {
      * 文件缓存
      */
     private Map<String, Element> fileCache;
-
 
     /**
      * 目录中的虚拟容器缓存
@@ -73,22 +81,50 @@ public class VirtualContainer implements Closeable {
      * @throws IllegalArgumentException 参数错误
      */
     public VirtualContainer(Path fullDir) throws IllegalArgumentException {
+        this(fullDir, ContainerType.FILE_SYSTEM);
+    }
+
+    public VirtualContainer(Path fullDir, ContainerType containerType) throws IllegalArgumentException {
         this();
-        if (fullDir == null) {
-            throw new IllegalArgumentException("完整路径(fullDir)为空");
-        }
-        // 目录不存在或不是一个目录
-        if (Files.notExists(fullDir) || !Files.isDirectory(fullDir)) {
-            try {
-                // 创建目录
-                fullDir = Files.createDirectories(fullDir);
-            } catch (IOException e) {
-                throw new RuntimeException("无法创建指定目录", e);
+        setContainer(fullDir, containerType, zipContainer);
+    }
+
+    public VirtualContainer(ContainerArgs containerArgs) throws IllegalArgumentException {
+        this();
+        setContainer(containerArgs.getFullDir(), containerArgs.getContainerType(), containerArgs.getZipContainer());
+    }
+
+    public void setContainer(Path fullDir, ContainerType containerType, ZipContainer zipContainer) {
+        this.containerType = containerType;
+        if (containerType == ContainerType.FILE_SYSTEM) {
+            if (fullDir == null) {
+                throw new IllegalArgumentException("完整路径(fullDir)为空");
+            }
+            // 目录不存在或不是一个目录
+            if (Files.notExists(fullDir) || !Files.isDirectory(fullDir)) {
+                try {
+                    // 创建目录
+                    fullDir = Files.createDirectories(fullDir);
+                } catch (IOException e) {
+                    throw new RuntimeException("无法创建指定目录", e);
+                }
+            }
+            this.fullPath = fullDir.toAbsolutePath().toString();
+            this.name = fullDir.getFileName().toString();
+        } else if (containerType == ContainerType.ZIP_MEMORY_FILE) {
+            if (fullDir != null) {
+                this.fullPath = fullDir.toString();
+            }
+            if (zipContainer != null) {
+                this.zipContainer = zipContainer;
+            } else {
+                try {
+                    this.zipContainer = new ZipContainer();
+                } catch (IOException e) {
+                    e.printStackTrace();
+                }
             }
         }
-        this.fullPath = fullDir.toAbsolutePath().toString();
-        this.name = fullDir.getFileName().toString();
-
     }
 
     /**
@@ -151,6 +187,23 @@ public class VirtualContainer implements Closeable {
         return this;
     }
 
+    public VirtualContainer putFile(String resourceFileName, byte[] resourceData) throws IOException {
+        if (resourceData == null || resourceFileName == null || resourceFileName.isEmpty()) {
+            return this;
+        }
+        Path target = Paths.get(fullPath, resourceFileName);
+        if (containerType == ContainerType.FILE_SYSTEM) {
+            // 写文件到指定目录.
+            FileOutputStream outputStream = new FileOutputStream(target.toAbsolutePath().toString());
+            outputStream.write(resourceData);
+            outputStream.close();
+        } else if (containerType == ContainerType.ZIP_MEMORY_FILE) {
+            zipContainer.addFile(target, resourceData);
+        }
+
+        return this;
+    }
+
     /**
      * 向虚拟容器加入对象
      *
@@ -189,10 +242,20 @@ public class VirtualContainer implements Closeable {
         if (element == null) {
             // 缓存中不存在，从文件目录中尝试读取
             Path file = getFile(fileName);
-            // 反序列化文件为对象
-            element = ElemCup.inject(file);
-            // 从文件加载元素，那么缓存该元素对象
-            fileCache.put(fileName, element);
+            if (containerType == ContainerType.FILE_SYSTEM) {
+                // 反序列化文件为对象
+                element = ElemCup.inject(file);
+            } else if (containerType == ContainerType.ZIP_MEMORY_FILE) {
+                byte[] fileData = zipContainer.readFile(file);
+                if (fileData == null) {
+                    throw new FileNotFoundException("无法在ZIP: " + fullPath + "中找到，文件 [ " + fileName + " ]");
+                }
+                element = ElemCup.inject(fileData);
+            }
+            if (element != null) {
+                // 从文件加载元素，那么缓存该元素对象
+                fileCache.put(fileName, element);
+            }
         }
         return element;
     }
@@ -209,12 +272,13 @@ public class VirtualContainer implements Closeable {
             throw new IllegalArgumentException("文件名为空");
         }
         Path res = Paths.get(fullPath, fileName);
-        if (Files.isDirectory(res) || Files.notExists(res)) {
-            throw new FileNotFoundException("无法在目录: " + fullPath + "中找到，文件 [ " + fileName + " ]");
+        if (containerType == ContainerType.FILE_SYSTEM) {
+            if (Files.isDirectory(res) || Files.notExists(res)) {
+                throw new FileNotFoundException("无法在目录: " + fullPath + "中找到，文件 [ " + fileName + " ]");
+            }
         }
         return res;
     }
-
 
     /**
      * 获取一个虚拟容器对象
@@ -223,11 +287,43 @@ public class VirtualContainer implements Closeable {
      * <p>
      * 如果容器不存在，那么创建一个新的对象
      *
-     * @param name   容器名称
-     * @param mapper 容器构造器引用
-     * @param <R>    容器子类
+     * @param name          容器名称
+     * @param containerType 容器类型
+     * @param zipContainer  容器对象
+     * @param mapper        容器构造器引用
+     * @param <R>           容器子类
      * @return 新建或已经存在的容器
      */
+    public <R extends VirtualContainer> R obtainContainer(String name, ContainerType containerType, ZipContainer zipContainer, Function<ContainerArgs, R> mapper) {
+        if (name == null || name.length() == 0) {
+            throw new IllegalArgumentException("容器名称（name）为空");
+        }
+        if (mapper == null) {
+            throw new IllegalArgumentException("容器构建对象（mapper）为空");
+        }
+        // 检查缓存
+        VirtualContainer target = dirCache.get(name);
+        if (target == null) {
+            Path p = Paths.get(name);
+            if (this.fullPath != null) {
+                p = Paths.get(this.fullPath, name);
+            }
+            // 如果目录不存在那么创建，如果已经存在那么就是加载
+            ContainerArgs containerArgs = new ContainerArgs();
+            containerArgs.setFullDir(p);
+            containerArgs.setContainerType(containerType);
+            containerArgs.setZipContainer(zipContainer);
+            R ct = mapper.apply(containerArgs);
+            // 设置父母路径
+            ct.setParent(this);
+            // 加入缓存中
+            dirCache.put(name, ct);
+            return ct;
+        } else {
+            return (R) target;
+        }
+    }
+
     public <R extends VirtualContainer> R obtainContainer(String name, Function<Path, R> mapper) {
         if (name == null || name.length() == 0) {
             throw new IllegalArgumentException("容器名称（name）为空");
@@ -238,7 +334,10 @@ public class VirtualContainer implements Closeable {
         // 检查缓存
         VirtualContainer target = dirCache.get(name);
         if (target == null) {
-            Path p = Paths.get(fullPath, name);
+            Path p = Paths.get(name);
+            if (this.fullPath != null) {
+                p = Paths.get(this.fullPath, name);
+            }
             // 如果目录不存在那么创建，如果已经存在那么就是加载
             R ct = mapper.apply(p);
             // 设置父母路径
@@ -261,16 +360,51 @@ public class VirtualContainer implements Closeable {
      * @throws FileNotFoundException 文件不存在
      */
     public <R extends VirtualContainer> R getContainer(String name, Function<Path, R> mapper) throws FileNotFoundException {
-        Path p = Paths.get(fullPath, name);
-        if (Files.notExists(p) || !Files.isDirectory(p)) {
-            throw new FileNotFoundException("容器内无法找名为：" + name + "目录");
+        Path p = Paths.get(name);
+        if (this.fullPath != null) {
+            p = Paths.get(this.fullPath, name);
         }
-
+        if (containerType == ContainerType.FILE_SYSTEM) {
+            if (Files.notExists(p) || !Files.isDirectory(p)) {
+                throw new FileNotFoundException("容器内无法找名为：" + name + "目录");
+            }
+        }
         // 检查缓存
         VirtualContainer target = dirCache.get(name);
         if (target == null) {
             // 调用指定构造器创建容器对象
             R ct = mapper.apply(p);
+            // 设置所属容器为创建者
+            ct.setParent(this);
+            // 设置容器.
+            ct.setContainer(p, containerType, zipContainer);
+            // 加入缓存中
+            dirCache.put(name, ct);
+            return ct;
+        } else {
+            return (R) target;
+        }
+    }
+
+    public <R extends VirtualContainer> R getContainerByContainerArgs(String name, Function<ContainerArgs, R> mapper) throws FileNotFoundException {
+        Path p = Paths.get(name);
+        if (this.fullPath != null) {
+            p = Paths.get(this.fullPath, name);
+        }
+        if (containerType == ContainerType.FILE_SYSTEM) {
+            if (Files.notExists(p) || !Files.isDirectory(p)) {
+                throw new FileNotFoundException("容器内无法找名为：" + name + "目录");
+            }
+        }
+        // 检查缓存
+        VirtualContainer target = dirCache.get(name);
+        if (target == null) {
+            // 调用指定构造器创建容器对象
+            ContainerArgs containerArgs = new ContainerArgs();
+            containerArgs.setFullDir(p);
+            containerArgs.setContainerType(this.containerType);
+            containerArgs.setZipContainer(this.zipContainer);
+            R ct = mapper.apply(containerArgs);
             // 设置所属容器为创建者
             ct.setParent(this);
             // 加入缓存中
@@ -339,16 +473,18 @@ public class VirtualContainer implements Closeable {
      * 删除整个虚拟容器
      */
     public void clean() {
-        try {
-            Path path = getContainerPath();
-            // 删除整个文件目录
-            if (Files.exists(path)) {
-                FileUtils.deleteDirectory(path.toFile());
+        if (containerType == ContainerType.FILE_SYSTEM) {
+            try {
+                Path path = getContainerPath();
+                // 删除整个文件目录
+                if (Files.exists(path)) {
+                    FileUtils.deleteDirectory(path.toFile());
+                }
+                this.fileCache.clear();
+                this.dirCache.clear();
+            } catch (Exception e) {
+                System.err.println("容器删除异常: " + e.getMessage());
             }
-            this.fileCache.clear();
-            this.dirCache.clear();
-        } catch (Exception e) {
-            System.err.println("容器删除异常: " + e.getMessage());
         }
     }
 
@@ -358,11 +494,25 @@ public class VirtualContainer implements Closeable {
      * @throws IOException 文件读写IO异常
      */
     public void flush() throws IOException {
-        // 刷新元素对象到指定目录
-        for (Map.Entry<String, Element> kv : fileCache.entrySet()) {
-            Path filePath = Paths.get(fullPath, kv.getKey());
-            // 序列化为文件
-            ElemCup.dump(kv.getValue(), filePath);
+        if (containerType == ContainerType.FILE_SYSTEM) {
+            // 刷新元素对象到指定目录
+            for (Map.Entry<String, Element> kv : fileCache.entrySet()) {
+                Path filePath = Paths.get(fullPath, kv.getKey());
+                //System.out.println(filePath.toString());
+                // 序列化为文件.
+                ElemCup.dump(kv.getValue(), filePath);
+            }
+        } else if (containerType == ContainerType.ZIP_MEMORY_FILE) {
+            for (Map.Entry<String, Element> kv : fileCache.entrySet()) {
+                Path filePath = Paths.get(kv.getKey());
+                if (this.fullPath != null) {
+                    filePath = Paths.get(this.fullPath, kv.getKey());
+                }
+                //System.out.println(filePath.toString());
+                // 序列化为内存对象.
+                byte[] fileData = ElemCup.dump(kv.getValue());
+                zipContainer.addFile(filePath, fileData);
+            }
         }
         // 递归的刷新容器中包含的其他容器
         for (VirtualContainer container : dirCache.values()) {
@@ -429,4 +579,5 @@ public class VirtualContainer implements Closeable {
         // 删除工作过程中存放于虚拟容器中的文件和目录
         flush();
     }
+
 }
